+++ MulliganMadness/Cards/JarOfDirt.cs
@@ -17,23 +17,7 @@ namespace MulliganMadness.Cards
 
         protected override GameObject GetCardArt() => CardArtFactory.Create("jarofdirt");
 
-        protected override CardInfoStat[] GetStats() => new[]
-        {
-            new CardInfoStat
-            {
-                positive = true,
-                stat = "Nulls",
-                amount = "Become treasures",
-                simepleAmount = CardInfoStat.SimpleAmount.notAssigned
-            },
-            new CardInfoStat
-            {
-                positive = false,
-                stat = "Disabled Nulls",
-                amount = "Unchanged",
-                simepleAmount = CardInfoStat.SimpleAmount.notAssigned
-            }
-        };
+        protected override CardInfoStat[] GetStats() => System.Array.Empty<CardInfoStat>();
 
         public override bool GetEnabled() => true;