+++ MulliganMadness/Curses/ForcedChoice.cs
@@ -16,15 +16,6 @@ namespace MulliganMadness.Curses
 
         protected override CardInfo.Rarity GetRarity() => CardInfo.Rarity.Common;
 
-        protected override CardInfoStat[] GetStats() => new[]
-        {
-            new CardInfoStat
-            {
-                positive = false,
-                stat = "Card pick",
-                amount = "Random",
-                simepleAmount = CardInfoStat.SimpleAmount.notAssigned
-            }
-        };
+        protected override CardInfoStat[] GetStats() => System.Array.Empty<CardInfoStat>();
     }
 }